feat: updating login activity logic, finished firebase implementation, handled some bugs
