+++ app/src/main/AndroidManifest.xml
@@ -49,7 +49,7 @@
             android:name=".SplashScreenActivity"
             android:exported="true"
             android:label="Recipe"
-            android:theme="@style/Theme.Recipe_App_Compose">
+            android:theme="@style/Theme.AppCompat.DayNight">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -60,20 +60,12 @@
         <activity
             android:name=".LoginActivity"
             android:exported="true"
-            android:theme="@style/Theme.Recipe_App_Compose" />
+            android:theme="@style/Theme.AppCompat.DayNight" />
 
         <activity
             android:name=".MainActivity"
             android:exported="true"
             android:theme="@style/Theme.Recipe_App_Compose" />
-        <receiver
-            android:name=".core.util.ConnectivityReceiver"
-            android:enabled="true"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.net.conn.CONNECTIVITY_CHANGE" />
-            </intent-filter>
-        </receiver>
 
     </application>
