--- app/src/main/java/com/example/recipe_app_compose/features/categories/presentation/view/LoginScreen.kt
@@ -1,234 +1,0 @@
-@file:OptIn(ExperimentalMaterial3Api::class)
-
-package com.example.recipe_app_compose.features.categories.presentation.view
-
-import android.app.Activity
-import android.content.Context
-import android.content.Intent
-import android.util.Patterns
-import android.widget.Toast
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.aspectRatio
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardActions
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Lock
-import androidx.compose.material.icons.filled.Person
-import androidx.compose.material3.Button
-import androidx.compose.material3.Checkbox
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusDirection
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalFocusManager
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import coil.ImageLoader
-import coil.compose.rememberAsyncImagePainter
-import com.example.recipe_app_compose.MainActivity
-import com.example.recipe_app_compose.R
-
-@Composable
-fun LoginForm() {
-    Surface {
-        var credentials by remember { mutableStateOf(Credentials()) }
-        val context = LocalContext.current
-
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(horizontal = 30.dp, vertical = 100.dp)
-        ) {
-            Image(painter = rememberAsyncImagePainter(
-                R.drawable.dining_two,
-                imageLoader = ImageLoader.Builder(context).crossfade(1000).build()),
-                contentDescription = "Image",
-                modifier = Modifier
-                    .height(150.dp)
-                    .aspectRatio(1f)
-            )
-            Spacer(modifier = Modifier.height(30.dp))
-            Text(
-                "Favorite Cuisines...",
-                style = MaterialTheme.typography.headlineMedium,
-                textAlign = TextAlign.Start,
-                modifier = Modifier.padding(bottom = 20.dp).fillMaxWidth()
-            )
-            LoginField(
-                value = credentials.email,
-                onChange = { data -> credentials = credentials.copy(email = data) },
-                modifier = Modifier.fillMaxWidth()
-            )
-            Spacer(modifier = Modifier.height(10.dp))
-            PasswordField(
-                value = credentials.pwd,
-                onChange = { data -> credentials = credentials.copy(pwd = data) },
-                submit = {
-                    if (!checkCredentials(credentials, context)) credentials = Credentials()
-                },
-                modifier = Modifier.fillMaxWidth()
-            )
-            Spacer(modifier = Modifier.height(10.dp))
-            LabeledCheckbox(
-                label = "Remember Me",
-                onCheckChanged = {
-                    credentials = credentials.copy(remember = !credentials.remember)
-                },
-                isChecked = credentials.remember
-            )
-            Spacer(modifier = Modifier.height(20.dp))
-            Button(
-                onClick = {
-                    if (!checkCredentials(credentials, context)) credentials = Credentials()
-                },
-                enabled = credentials.isNotEmpty(),
-                shape = RoundedCornerShape(5.dp),
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                Text("Login")
-            }
-        }
-    }
-}
-
-fun checkCredentials(creds: Credentials, context: Context): Boolean {
-    if (creds.isNotEmpty() && (creds.email.matches(Patterns.EMAIL_ADDRESS.toRegex()))) {
-        context.startActivity(Intent(context, MainActivity::class.java))
-        (context as Activity).finish()
-        return true
-    } else {
-        Toast.makeText(context, "Wrong Credentials", Toast.LENGTH_SHORT).show()
-        return false
-    }
-}
-
-data class Credentials(
-    var email: String = "",
-    var pwd: String = "",
-    var remember: Boolean = false
-) {
-    fun isNotEmpty(): Boolean {
-        return email.isNotEmpty() && pwd.isNotEmpty()
-    }
-}
-
-
-@Composable
-fun LabeledCheckbox(
-    label: String,
-    onCheckChanged: () -> Unit,
-    isChecked: Boolean
-) {
-
-    Row(
-        Modifier
-            .clickable(
-                onClick = onCheckChanged
-            )
-            .padding(4.dp)
-    ) {
-        Checkbox(checked = isChecked, onCheckedChange = null)
-        Spacer(Modifier.size(6.dp))
-        Text(label)
-    }
-}
-
-@Composable
-fun LoginField(
-    value: String,
-    onChange: (String) -> Unit,
-    modifier: Modifier = Modifier,
-    label: String = "Login",
-    placeholder: String = "Enter your Login"
-) {
-
-    val focusManager = LocalFocusManager.current
-    val leadingIcon = @Composable {
-        Icon(
-            Icons.Default.Person,
-            contentDescription = "",
-            tint = MaterialTheme.colorScheme.primary
-        )
-    }
-
-    OutlinedTextField(
-        value = value,
-        onValueChange = onChange,
-        modifier = modifier,
-        leadingIcon = leadingIcon,
-        keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next),
-        keyboardActions = KeyboardActions(
-            onNext = { focusManager.moveFocus(FocusDirection.Down) }
-        ),
-        placeholder = { Text(placeholder) },
-        label = { Text(label) },
-        singleLine = true,
-        visualTransformation = VisualTransformation.None,
-        shape = RoundedCornerShape(20.dp)
-    )
-}
-
-@Composable
-fun PasswordField(
-    value: String,
-    onChange: (String) -> Unit,
-    submit: () -> Unit,
-    modifier: Modifier = Modifier,
-    label: String = "Password",
-    placeholder: String = "Enter your Password"
-) {
-
-    val isPasswordVisible by remember { mutableStateOf(false) }
-
-    val leadingIcon = @Composable {
-        Icon(
-            Icons.Default.Lock,
-            contentDescription = "",
-            tint = MaterialTheme.colorScheme.primary
-        )
-    }
-    OutlinedTextField(
-        value = value,
-        onValueChange = onChange,
-        modifier = modifier,
-        leadingIcon = leadingIcon,
-        keyboardOptions = KeyboardOptions(
-            imeAction = ImeAction.Done,
-            keyboardType = KeyboardType.Password
-        ),
-        keyboardActions = KeyboardActions(
-            onDone = { submit() }
-        ),
-        placeholder = { Text(placeholder) },
-        label = { Text(label) },
-        singleLine = true,
-        visualTransformation = if (isPasswordVisible) VisualTransformation.None else PasswordVisualTransformation(),
-        shape = RoundedCornerShape(20.dp)
-    )
-}
